This change is to cover a possible schenario where the service starting docker-wait-any ( swss) or one of the dependent service exists immediatly after start_peer_dependent_services in swss.sh script. Though we wait for 60 secs to make sure the service comes up, assume it doesn't come up.
In this scenario, when we call docker-wait-any from swss.sh, it will come and could get stuck in the while loop in this
function wait_for_container() here in this script. This while loop was introduced earlier to handle the case
when a dependent service eg: teamd goes down and has warm restart enabled.

So moving this while loop wait to inside the warm/fast boot check.